# cmd/webserver/pages/api/generated/openapi.json
{
  "components": {
    "parameters": {
      "limit-param": {
        "in": "query",
        "name": "limit",
        "schema": {
          "default": 10,
          "maximum": 100,
          "minimum": 1,
          "type": "integer"
        }
      },
      "offset-param": {
        "in": "query",
        "name": "offset",
        "schema": {
          "default": 0,
          "minimum": 0,
          "type": "integer"
        }
      }
    },
    "responses": {
...
      "apps-response": {
        "content": {
          "application/json": {
            "schema": {
              "description": "List of apps, with pagination",
              "properties": {
                "apps": {
                  "items": {
                    "$ref": "#/components/schemas/app-schema"
                  },
                  "type": "array"
                },
                "pagination": {
                  "$ref": "#/components/schemas/pagination-schema"
                }
              },
              "required": [
                "pagination",
                "apps"
              ]
            }
          }
        },
        "description": "List of apps"
      },
...
      "players-response": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "pagination": {
                  "$ref": "#/components/schemas/pagination-schema"
                },
                "players": {
                  "items": {
                    "$ref": "#/components/schemas/player-schema"
                  },
                  "type": "array"
                }
              },
              "required": [
                "pagination",
                "players"
              ]
            }
          }
        },
        "description": "List of players"
      }
    },
    "schemas": {
      "app-schema": {
        "properties": {
          "categories": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "developers": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "genres": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "id": {
            "type": "integer"
          },
          "metacritic_score": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "players_max": {
            "type": "integer"
          },
          "players_week_avg": {
            "format": "double",
            "type": "number"
          },
          "players_week_max": {
            "type": "integer"
          },
          "prices": {
            "items": {
              "properties": {
                "currency": {
                  "type": "string"
                },
                "discountPercent": {
                  "format": "int32",
                  "type": "integer"
                },
                "final": {
                  "format": "int32",
                  "type": "integer"
                },
                "individual": {
                  "format": "int32",
                  "type": "integer"
                },
                "initial": {
                  "format": "int32",
                  "type": "integer"
                }
              },
              "required": [
                "currency",
                "initial",
                "final",
                "discountPercent",
                "individual"
              ]
            },
            "type": "array"
          },
          "publishers": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          },
          "release_date": {
            "format": "int64",
            "type": "integer"
          },
          "reviews_negative": {
            "type": "integer"
          },
          "reviews_positive": {
            "type": "integer"
          },
          "reviews_score": {
            "format": "double",
            "type": "number"
          },
          "tags": {
            "items": {
              "type": "integer"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "tags",
          "genres",
          "categories",
          "developers",
          "publishers",
          "prices",
          "players_max",
          "players_week_max",
          "players_week_avg",
          "release_date",
          "reviews_positive",
          "reviews_negative",
          "reviews_score",
          "metacritic_score"
        ]
      },
...
      "pagination-schema": {
        "properties": {
          "limit": {
            "format": "int64",
            "type": "integer"
          },
          "offset": {
            "format": "int64",
            "type": "integer"
          },
          "pagesCurrent": {
            "type": "integer"
          },
          "pagesTotal": {
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "offset",
          "limit",
          "total",
          "pagesTotal",
          "pagesCurrent"
        ]
      },
      "player-schema": {
        "properties": {
          "avatar": {
            "type": "string"
          },
          "badges": {
            "type": "integer"
          },
          "comments": {
            "type": "integer"
          },
          "continent": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "friends": {
            "type": "integer"
          },
          "games": {
            "type": "integer"
          },
          "groups": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "level": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "playtime": {
            "type": "integer"
          },
          "state": {
            "type": "string"
          },
          "vanity_url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "avatar",
          "badges",
          "comments",
          "friends",
          "games",
          "groups",
          "level",
          "playtime",
          "country",
          "continent",
          "state",
          "vanity_url"
        ]
      },
...
    },
    "securitySchemes": {
      "key-header": {
        "in": "header",
        "name": "key",
        "type": "apiKey"
      },
      "key-query": {
        "in": "query",
        "name": "key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "email": "jimeagle@gmail.com",
      "name": "Jleagle",
      "url": "https://gamedb.online/contact"
    },
    "title": "Game DB API",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
    "/games": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "pattern": "^[0-9A-Z]{20}$",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/offset-param"
          },
          {
            "$ref": "#/components/parameters/limit-param"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "default": "asc",
              "enum": [
                [
                  "asc",
                  "desc"
                ]
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ids",
            "schema": {
              "items": {
                "type": "integer"
              },
              "maxItems": 100,
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "items": {
                "type": "integer"
              },
              "maxItems": 10,
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "genres",
            "schema": {
              "items": {
                "type": "integer"
              },
              "maxItems": 10,
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "categories",
            "schema": {
              "items": {
                "type": "integer"
              },
              "maxItems": 10,
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "developers",
            "schema": {
              "items": {
                "type": "integer"
              },
              "maxItems": 10,
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "publishers",
            "schema": {
              "items": {
                "type": "integer"
              },
              "maxItems": 10,
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "platforms",
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 3,
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/apps-response"
          }
        },
        "summary": "List Apps"
      }
    },
...
    "/players": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "pattern": "^[0-9A-Z]{20}$",
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/offset-param"
          },
          {
            "$ref": "#/components/parameters/limit-param"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "default": "id",
              "enum": [
                "id",
                "level",
                "badges",
                "games",
                "time",
                "friends",
                "comments"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "default": "desc",
              "enum": [
                [
                  "asc",
                  "desc"
                ]
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "continent",
            "schema": {
              "items": {
                "maxLength": 2,
                "type": "string"
              },
              "maxItems": 3,
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "country",
            "schema": {
              "items": {
                "maxLength": 2,
                "type": "string"
              },
              "maxItems": 3,
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/players-response"
          }
        },
        "summary": "List Players"
      }
    },
...
  },
  "security": [
    {
      "key-header": []
    },
    {
      "key-query": []
    }
  ],
  "servers": [
    {
      "url": "https://gamedb.online/api"
    }
  ]
}